server/pyproject.toml
[tool.poetry]
name = "server_config"
version = "0.1.0"
description = "Backend for library management with Django"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
psycopg2-binary = "^2.9.9"
django-celery-results = "^2.5.1"
celery = "^5.4.0"
django-prometheus = "^2.3.1"
python-jose = "^3.3.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
gunicorn = "^22.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"